+++ changelog.md
@@ -7,10 +7,27 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.1.1] - 2018-11-02
+
+### Added
+
+- Added the decoded token to error handler when it is available.
+
+### Changed
+
+- Changed `JWTManager` from default export to named export.
+
+- Made the argument for `JWTManager#invalidate` optional.
+
+### Removed
+
+- Removed intallation with tag 'latest' for GitHub in readme.
+
 ## 0.1.0 - 2018-11-01
 
 ### Added
 
 - Initial release
 
-[Unreleased]: https://github.com/revam/node-jwt-manager/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/revam/node-jwt-manager/compare/v0.1.1...HEAD
+[0.1.1]: https://github.com/revam/node-jwt-manager/compare/v0.1.0...v0.1.1